\name{fpt}
\alias{fpt}
\title{
C++ Function to be Built Using \code{GQD.TIpassage()}
}
\description{
This is a peripheral function that is created in C++ using \code{GQD.TIpassage()}. This is a non-critical function that is re-defined internally when the need arises. Included here for documentation purposes.
}
\usage{
fpt(res, res2, N, delt)
}

\arguments{
  \item{res}{Solution to a system of ODES.
}
  \item{res2}{Solution to a system of ODES.
}
  \item{N}{Integer giving number of time nodes.
}
  \item{delt}{Step size.
}
}
\value{
Varies by model.
}

\author{
Etienne A.D. Pienaar: \email{etiennead@gmail.com}
}

