\name{priors}
\alias{priors}

\title{
Define Prior Densities on a Parameter Set on a GQD Model.
}
\description{
\code{priors()} can be used to define prior densities on the parameters of a GQD model 
}
\usage{
priors(theta)
}
\arguments{
  \item{theta}{Vector of parameters.
}
}

\value{
Returns a single numeric value. 
}


\author{
Etienne A.D. Pienaar: \email{etiennead@gmail.com}
}