\name{solverFPT}
\alias{solverFPT}

\title{
C++ Function to be Built Using \code{GQD.TIpassage()}
}
\description{
This is a peripheral function that is created in C++ using \code{GQD.TIpassage()}. This is a non-critical function that is re-defined internally when the need arises. Included here for documentation purposes.
}
\usage{
solverFPT(Xs, Bt, theta, N, delt, tt)
}

\arguments{
  \item{Xs}{
Initial value.
}
  \item{Bt}{
  Barrier.
}
  \item{theta}{
  Parameter vector.
}
  \item{N}{
Integer indicating number of iterations.
}
  \item{delt}{
Step size.
}
  \item{tt}{
A time sequence.
}
}
\value{
Varies by model.
}

\author{
Etienne A.D. Pienaar: \email{etiennead@gmail.com}
}
